requirements.txt
wheel==0.43.0
tqdm==4.66.4
pytorch_lightning==2.3.0
black==24.4.2
ruff==0.4.8
hydra-core==1.3.2
jaxtyping==0.2.30
beartype==0.18.5
wandb==0.17.1
einops==0.8.0
colorama==0.4.6
scikit-image==0.23.2
colorspacious==1.1.2
matplotlib==3.5.2
moviepy==1.0.3
imageio==2.34.1
git+https://github.com/dcharatan/diff-gaussian-rasterization-modified
timm==1.0.3
dacite==1.8.1
lpips==0.1.4
e3nn==0.5.1
plyfile==1.0.3
tabulate==0.9.0
svg.py==1.4.3
opencv-python==4.6.0.66
sk-video==1.1.10
